requirements.txt
Django==3.1.1
gunicorn==20.0.4
djangorestframework>=1.0
requests>=1.0